+++ src/app.js
@@ -1,19 +1,28 @@
 require('dotenv').config();
 const express = require('express');
-const indexRouter = require('./routes/index')
+const createRouter = require('./routes/create/index');
+const readRouter = require('./routes/read/index');
+const updateRouter = require('./routes/update/index');
+const deleteRouter = require('./routes/delete/index');
 const app = express();
 
 
 app.use(express.json());
 
-app.use('/', indexRouter)
+app.use('/', (req, res, next) => {
+    if (req.originalUrl === '/') {
+        res.json({ message: 'Service is running!'});
+        return;
+    }
+    next();
+});
 
-// app.use('/', (req, res, next) => {
-//     if (req.originalUrl === '/') {
-//         res.send('Service is running!');
-//         return;
-//     }
-//     next();
-// });
+app.use('/',
+    [
+        createRouter,
+        readRouter,
+        updateRouter,
+        deleteRouter
+    ]);
 
 module.exports = app;
