+++ test/app.integration.spec.js
@@ -3,7 +3,7 @@ const app = require('../src/app');
 const connection = require('../src/connection');
 
 describe('Test routes', () => {
-    //beforeEach(done => connection.query('TRUNCATE anime_list', done));
+    beforeEach(done => connection.query('DELETE FROM anime_list WHERE title = "Neon Genesis Evangelion"', done));
     it('GET / sends "Service is running!" as json', (done) => {
         request(app)
             .get('/')
@@ -16,25 +16,13 @@ describe('Test routes', () => {
             });
     });
 
-    // it('GET /anime retrives all of the data from a table', (done) => {
-    //     request(app)
-    //         .get('/anime')
-    //         .expect(200)
-    //         .expect('Content-Type', /json/)
-    //         .then(response => {
-    //             const expected = [{}];
-    //             expect(response.body).toEqual(expected);
-    //             done();
-    //         });
-    // });
-
     it('GET /anime/:id retrives exact feilds from a table', (done) => {
         request(app)
             .get('/anime/1')
             .expect(200)
             .expect('Content-Type', /json/)
             .then(response => {
-                const expected = { id: expect.any(Number), title: 'Neon Genesis Evangelion', year: "1995-10-02T22:00:00.000Z", episodes: 26, is_checked: 1 };
+                const expected = { id: 1, title: expect.any(String), year: "2003-12-31T22:00:00.000Z", episodes: expect.any(Number), is_checked: 1 };
                 expect(response.body).toEqual(expected);
                 done();
             });
@@ -42,7 +30,7 @@ describe('Test routes', () => {
 
     it('GET /anime/:id - error: "Anime not found"', (done) => {
         request(app)
-            .get('/anime/10')
+            .get('/anime/101')
             .expect(404)
             .expect('Content-Type', /json/)
             .then(response => {
@@ -56,10 +44,10 @@ describe('Test routes', () => {
         request(app)
         .post('/anime')
           .expect(201)
-          .send({ title: 'Neon Genesis Evangelion', year: "1995-10-02T22:00:00.000Z", episodes: 26, is_checked: 1 })
+          .send({ title: 'Neon Genesis Evangelion', year: "1995-10-02", episodes: 26, is_checked: true })
           .expect('Content-Type', /json/)
           .then(response => {
-            const expected = { id: expect.any(Number), title: 'Neon Genesis Evangelion', year: expect.to.be(Date), episodes: 26, is_checked: 1 };
+            const expected = { id: expect.any(Number), title: 'Neon Genesis Evangelion', year: "1995-10-01T22:00:00.000Z", episodes: 26, is_checked: 1 };
             expect(response.body).toEqual(expected);
             done();
           })
